Split out components into individual files
